refactor(test): 修复 golangci-lint 全部 10 个问题
