+++ migrate/v2/v160_logid0_test.go
@@ -1,17 +1,18 @@
-package v2
+package v2_test
 
 import (
 	"testing"
 
-	"sealdice-core/migrate/v2/v160"
+	v160 "sealdice-core/migrate/v2/v160"
+	"sealdice-core/migrate/v2/v2test"
 	"sealdice-core/utils/constant"
 )
 
 func TestV160LogIDZeroClean_DeletesAndRecounts(t *testing.T) {
-	op, _ := newTestSQLiteEngine(t)
+	op, _ := v2test.NewTestSQLiteEngine(t)
 	logDB := op.GetLogDB(constant.WRITE)
 
-	mustExec(t, logDB, `CREATE TABLE logs (
+	v2test.MustExec(t, logDB, `CREATE TABLE logs (
 		id INTEGER PRIMARY KEY AUTOINCREMENT,
 		name TEXT,
 		group_id TEXT,
@@ -19,20 +20,19 @@ func TestV160LogIDZeroClean_DeletesAndRecounts(t *testing.T) {
 		updated_at INTEGER,
 		size INTEGER
 	)`)
-	mustExec(t, logDB, `CREATE TABLE log_items (
+	v2test.MustExec(t, logDB, `CREATE TABLE log_items (
 		id INTEGER PRIMARY KEY AUTOINCREMENT,
 		log_id INTEGER,
 		group_id TEXT,
 		removed INTEGER
 	)`)
-	mustExec(t, logDB, `INSERT INTO logs (id, name, size) VALUES (1, 'a', 99), (0, NULL, 0)`)
-	mustExec(t, logDB, `INSERT INTO log_items (log_id, removed) VALUES (1, NULL), (1, NULL), (0, NULL), (0, NULL)`)
+	v2test.MustExec(t, logDB, `INSERT INTO logs (id, name, size) VALUES (1, 'a', 99), (0, NULL, 0)`)
+	v2test.MustExec(t, logDB, `INSERT INTO log_items (log_id, removed) VALUES (1, NULL), (1, NULL), (0, NULL), (0, NULL)`)
 
-	if err := v160.V160LogIDZeroCleanMigrate(op, silentLogf); err != nil {
+	if err := v160.V160LogIDZeroCleanMigrate(op, v2test.SilentLogf); err != nil {
 		t.Fatalf("V160LogIDZeroCleanMigrate 失败: %v", err)
 	}
 
-	// id=0 的日志与 log_id=0 的条目应被删除
 	var logZero, itemLogZero int64
 	logDB.Raw("SELECT COUNT(1) FROM logs WHERE id = 0").Scan(&logZero)
 	logDB.Raw("SELECT COUNT(1) FROM log_items WHERE log_id = 0").Scan(&itemLogZero)
@@ -40,41 +40,35 @@ func TestV160LogIDZeroClean_DeletesAndRecounts(t *testing.T) {
 		t.Fatalf("log_id=0 清理不彻底: logs.id=0=%d, log_items.log_id=0=%d", logZero, itemLogZero)
 	}
 
-	// log 1 的 size 应被重算为 2（原本被设为脏值 99）
-	sizes := scanLogSizes(t, logDB)
+	sizes := v2test.ScanLogSizes(t, logDB)
 	if sizes[1] != 2 {
 		t.Fatalf("清理后 log 1 的 size 期望 2，实际 %d", sizes[1])
 	}
 }
 
-// 验证：当 size 列不存在（V150 失误遗留）时，008 迁移不会因重算失败而报错。
-// 此时清理仍会完成，size 列由 010 迁移负责补建与重算。
 func TestV160LogIDZeroClean_TolerantOfMissingSizeColumn(t *testing.T) {
-	op, _ := newTestSQLiteEngine(t)
+	op, _ := v2test.NewTestSQLiteEngine(t)
 	logDB := op.GetLogDB(constant.WRITE)
 
-	// 注意：logs 表没有 size 列
-	mustExec(t, logDB, `CREATE TABLE logs (
+	v2test.MustExec(t, logDB, `CREATE TABLE logs (
 		id INTEGER PRIMARY KEY AUTOINCREMENT,
 		name TEXT,
 		group_id TEXT,
 		created_at INTEGER,
 		updated_at INTEGER
 	)`)
-	mustExec(t, logDB, `CREATE TABLE log_items (
+	v2test.MustExec(t, logDB, `CREATE TABLE log_items (
 		id INTEGER PRIMARY KEY AUTOINCREMENT,
 		log_id INTEGER,
 		removed INTEGER
 	)`)
-	mustExec(t, logDB, `INSERT INTO logs (id, name) VALUES (1, 'a'), (0, NULL)`)
-	mustExec(t, logDB, `INSERT INTO log_items (log_id, removed) VALUES (1, NULL), (0, NULL)`)
+	v2test.MustExec(t, logDB, `INSERT INTO logs (id, name) VALUES (1, 'a'), (0, NULL)`)
+	v2test.MustExec(t, logDB, `INSERT INTO log_items (log_id, removed) VALUES (1, NULL), (0, NULL)`)
 
-	// 关键断言：不应报错
-	if err := v160.V160LogIDZeroCleanMigrate(op, silentLogf); err != nil {
+	if err := v160.V160LogIDZeroCleanMigrate(op, v2test.SilentLogf); err != nil {
 		t.Fatalf("size 列缺失时，008 迁移不应报错: %v", err)
 	}
 
-	// 清理仍应生效
 	var logZero, itemLogZero int64
 	logDB.Raw("SELECT COUNT(1) FROM logs WHERE id = 0").Scan(&logZero)
 	logDB.Raw("SELECT COUNT(1) FROM log_items WHERE log_id = 0").Scan(&itemLogZero)
@@ -84,23 +78,23 @@ func TestV160LogIDZeroClean_TolerantOfMissingSizeColumn(t *testing.T) {
 }
 
 func TestV160LogIDZeroClean_NothingToDoIsNoOp(t *testing.T) {
-	op, _ := newTestSQLiteEngine(t)
+	op, _ := v2test.NewTestSQLiteEngine(t)
 	logDB := op.GetLogDB(constant.WRITE)
 
-	mustExec(t, logDB, `CREATE TABLE logs (
+	v2test.MustExec(t, logDB, `CREATE TABLE logs (
 		id INTEGER PRIMARY KEY AUTOINCREMENT, name TEXT, group_id TEXT,
 		created_at INTEGER, updated_at INTEGER, size INTEGER
 	)`)
-	mustExec(t, logDB, `CREATE TABLE log_items (
+	v2test.MustExec(t, logDB, `CREATE TABLE log_items (
 		id INTEGER PRIMARY KEY AUTOINCREMENT, log_id INTEGER, removed INTEGER
 	)`)
-	mustExec(t, logDB, `INSERT INTO logs (id, name, size) VALUES (1, 'a', 1)`)
-	mustExec(t, logDB, `INSERT INTO log_items (log_id, removed) VALUES (1, NULL)`)
+	v2test.MustExec(t, logDB, `INSERT INTO logs (id, name, size) VALUES (1, 'a', 1)`)
+	v2test.MustExec(t, logDB, `INSERT INTO log_items (log_id, removed) VALUES (1, NULL)`)
 
-	if err := v160.V160LogIDZeroCleanMigrate(op, silentLogf); err != nil {
+	if err := v160.V160LogIDZeroCleanMigrate(op, v2test.SilentLogf); err != nil {
 		t.Fatalf("无 log_id=0 数据时不应报错: %v", err)
 	}
-	sizes := scanLogSizes(t, logDB)
+	sizes := v2test.ScanLogSizes(t, logDB)
 	if sizes[1] != 1 {
 		t.Fatalf("log 1 的 size 应保持不变为 1，实际 %d", sizes[1])
 	}
